// src/app/Models/Factuprint/products.ts
export interface Products {
    idArticulo: number;
    clave: string;
    nombre: string;
    precioVenta: number;
    precioCompra: number;
    unidadVenta: string;
    unidadCompra: string;
    relacion: number;
    idImp1: number;
    idImp2: number;
    idRet1: number;
    idRet2: number;
    existencia: number;
    observaciones: string;
    neto: number;
    netoC: number; 
    inventariable: number;
    costo: number;
    lotes: number;
    series: number;
    precioSug: number;
    oferta: string;
    promocion: string;
    impCig: number;
    color: number;
    precioLista: number;
    condiciones: string;
    utilidad: number;
    alterna: string;
    kit: number;
    dpc: number;
    dpv: number;
    reorden: number;
    maximo: number;
    kitSuelto: number;
    idClaseMultiple: number;
    prcFix: number;
    localiza: string;
}